requirements.txt
gradio==4.28.3
matplotlib==3.8.4
numpy==1.26.4
Pillow==10.3.0
tensorflow==2.16.1
tensorflow_hub==0.16.1
